Rename Comments -> CommentsActivity
This will make it easier to distinguish between CommentsActivity and the
soon-to-be created CommentsFragment.

As far as I can tell, this class name is used purely internally (it has
no intent mappings, for example), so we should be free to change it if
we wish.  A more experienced Android developer might be able to explain
whether or not this is actually a good idea.