Split CommentsFragment out of CommentsActivity
We want to display the comments to the right of the list of news
articles, but only when the screen is large enough.  But we don't want
to tear the application apart and spend the next two days putting it
back together.  So we're going to tackle this step-by-step, beginning
with creating a CommentsFragment that we can reuse in multiple
activities.

This code is heavily inspired by the following tutorial:

http://android-developers.blogspot.com/2011/02/android-30-fragments-api.html

This patch splits CommentsActivity in half, moving most of the code to
CommentsFragment.  It turns what's left of CommentsActivity into a thin
wrapper around CommentsFragment.  There are a few new APIs to allow the
two parts to communicate.