Action Bar: Centralize "home" handling and clear activity stack
We want to centralize the code which implements the "home" button.  This
involves more code duplication than I'd like, because Preferences
inherits from PreferencesActivity, and we therefore can't make it
inherit from the new HNActivity.

We also set Intent.FLAG_ACTIVITY_CLEAR_TOP so that when we click "home",
we pop our activity stack back to Main, instead of adding Main at the
top of the stack.