WebView: Display horizontal progress bar @ bottom during load
Before, some pages would hang forever, because we didn't show the page
until everything had loaded.  This patch switches to a horizontal
progress bar at the bottom of the screen and displays rendered content
immediately.

We should still customize this drawable a bit, but this fixes the most
urgent problems.