CommentsFragment: Stop accidentally calling main thread handler
We can't create handlers for nested fragments, because we can only
have one handler for the GUI thread (or so I understand it).  So
we need to overhaul CommentsFragment to stop accidentally triggering
the parent thread's handler when it means to trigger its own.

We also remove the "reload everything when upvoting a comment"
behavior, because it annoys me hugely.