Convert Handler subclasses to use runOnUiThread instead
The Handler subclasses in my new design were a bit messy, and they make
it hard to see the second half of any function that needs to run in the
UI thread.  So I'm converting all of them over to runOnUiThread calls.

This will make it marginally easier to attack some fragment transaction
problems in the next patch or two.