Massively overhaul fragment transactions, fixing several bugs
Before, we tended to create FragmentTransactions deep in the
call-chain.  This made it impossible to deal with the ActionBar tab API
correctly, because it passes us a pre-built transaction we are supposed
to use.

There are three major behavioral changes here:

1) Do not create WelcomeFragment on startup.  This seems to fix the
   "two stacked fragments" bug that I was seeing.  Instead, we rely
   on either the default tab click (for MainActivity) or the initial
   refreshNews call (for SubmissionsActivity) to reset this
   appropriately.

2) Use provided FragmentTransaction when responding to tab clicks.  This
   is certainly cleaner than ignoring it and trying to run our own
   transaction.

3) Reset detailsFragment to WelcomeFragment immediately when scheduling
   a refresh.  This helps support (2), and is once again a cleaner
   approach.

With any luck, we won't see the doubly-stacked fragments again.