tests: Generate a test project with a single, empty test
We have an annoying lifecycle bug that we'd like to track down, which
can theoretically be done using tests.  We also want to add tests for
new features and for any future regressions that we encounter, because
(a) I like test-driven development, and (b) hand-testing the application
is getting a bit tedious.